04 auth

+++ controllers/contactsControllers.js
@@ -3,13 +3,15 @@ import HttpError from "../helpers/HttpError.js";
 import ctrlWrapper from "../helpers/ctrlWrapper.js";
 
 export const getAllContacts = async (req, res) => {
-  const result = await contactsService.listContacts();
+  const { id } = req.user;
+  const result = await contactsService.listContacts({ owner: id });
   res.json(result);
 };
 
 export const getOneContact = async (req, res) => {
   const { id } = req.params;
-  const result = await contactsService.getContactById(id);
+  const { id: owner } = req.user;
+  const result = await contactsService.getContact({ id, owner });
   if (!result) {
     throw HttpError(404, "Not found");
   }
@@ -18,7 +20,8 @@ export const getOneContact = async (req, res) => {
 
 export const deleteContact = async (req, res) => {
   const { id } = req.params;
-  const result = await contactsService.removeContact(id);
+  const { id: owner } = req.user;
+  const result = await contactsService.removeContact({ id, owner });
   if (!result) {
     throw HttpError(404, "Not found");
   }
@@ -26,13 +29,15 @@ export const deleteContact = async (req, res) => {
 };
 
 export const createContact = async (req, res) => {
-  const result = await contactsService.addContact(req.body);
+  const { id } = req.user;
+  const result = await contactsService.addContact({ ...req.body, owner: id });
   res.status(201).json(result);
 };
 
 export const updateContact = async (req, res) => {
   const { id } = req.params;
-  const result = await contactsService.updateContactById(id, req.body);
+  const { id: owner } = req.user;
+  const result = await contactsService.updateContact({ id, owner }, req.body);
   if (!result) {
     throw HttpError(404, "Not found");
   }
@@ -42,7 +47,11 @@ export const updateContact = async (req, res) => {
 export const updateStatusContact = async (req, res) => {
   const { id } = req.params;
   const { favorite } = req.body;
-  const result = await contactsService.updateContactById(id, { favorite });
+  const { id: owner } = req.user;
+  const result = await contactsService.updateContact(
+    { id, owner },
+    { favorite }
+  );
   if (!result) {
     throw HttpError(404, "Not found");
   }
